requirements.txt
fastapi
uvicorn
redis
python-dotenv
requests